refactor(cdp): extract shared attachToPageTarget helper
Deduplicate identical CDP page-attach logic from BrowserStrategy and
CdpStorageExtractor into a single exported function in cdp-ws.ts.
